+++ app/Services/WsService.ts
@@ -1,6 +1,7 @@
 import { Server } from 'socket.io';
 import Redis from '@ioc:Adonis/Addons/Redis';
 import AdonisServer from '@ioc:Adonis/Core/Server';
+import { DateTime } from 'luxon';
 
 /**
  * Controllers
@@ -11,7 +12,13 @@ import OnlineController from 'App/Controllers/Ws/OnlineController';
  * Middlewares
  */
 import AuthMiddleware from 'App/Middleware/Ws/Auth';
+
+/**
+ * Stores
+ */
 import RedisSessionStore from 'App/Store/SessionStore';
+import RedisMessageStore from 'App/Store/MessageStore';
+import RedisNotificationStore from 'App/Store/NotificationStore';
 
 export interface ServerToClientEvents {
   /**
@@ -19,7 +26,18 @@ export interface ServerToClientEvents {
    */
   'user:session': (data: { sessionId: string | undefined; userId: string | undefined }) => void;
   'user:connected': (data: { userId: string; userName: string }) => void;
-  'user:online': (data: number) => void;
+  'user:online': (data: [string, { userId: string; userName: string }][]) => void;
+
+  /**
+   * Chat
+   */
+  'chat:message': (data: {
+    content: string;
+    time: string;
+    from: string;
+    to: string;
+    notificationId?: string | undefined;
+  }) => void;
 }
 
 export interface ClientToServerEvents {
@@ -28,6 +46,16 @@ export interface ClientToServerEvents {
    */
   'user:logout': () => void;
   'user:status': () => void;
+
+  /**
+   * Chat
+   */
+  'chat:message': (data: { content: string; to: string }) => void;
+
+  /**
+   * Notification
+   */
+  'notification:read': (data: { userId: string; ids: string[] }) => void;
 }
 
 export interface InterServerEvents {}
@@ -43,6 +71,10 @@ class WsService {
 
   public sessionStore: RedisSessionStore;
 
+  public messageStore: RedisMessageStore;
+
+  public notificationStore: RedisNotificationStore;
+
   public booted = false;
 
   public boot() {
@@ -57,24 +89,68 @@ class WsService {
         credentials: true,
       },
     });
-    this.sessionStore = new RedisSessionStore(Redis.connection('session'));
 
-    this.setupMiddlewares();
-    this.listen();
+    /**
+     * Stores
+     */
+    this.sessionStore = new RedisSessionStore(Redis.connection('session'));
+    this.messageStore = new RedisMessageStore(Redis.connection('message'));
+    this.notificationStore = new RedisNotificationStore(Redis.connection('notification'));
+
+    /**
+     * Clean up sessions before start
+     */
+    this.sessionStore.destroyAllSessions().then(() => {
+      /**
+       * Middlewares
+       */
+      this.setupMiddlewares();
+
+      /**
+       * On connection socket
+       */
+      this.listen();
+    });
   }
 
   private listen() {
-    this.io.on('connection', socket => {
+    this.io.on('connection', async socket => {
       // eslint-disable-next-line @typescript-eslint/no-unused-vars
       const onlineController = new OnlineController(socket, this.io, this.sessionStore);
 
       const { sessionId, userId, userName } = socket.data;
       const isExistSocketData = !!(sessionId && userId && userName);
 
+      if (isExistSocketData) {
+        socket.join(userId);
+      }
+
       /**
        * Send session to client
        */
       socket.emit('user:session', { sessionId: socket.data.sessionId, userId: socket.data.userId });
+
+      socket.on('chat:message', async ({ content, to }) => {
+        if (isExistSocketData) {
+          const message = {
+            content,
+            from: userId,
+            to,
+            time: DateTime.now().toISO(),
+          };
+
+          await this.messageStore.add(userId, to, message);
+
+          if (userId !== to) {
+            const notification = await this.notificationStore.add(to, `You received a new private message`, 'MESSAGE');
+            this.io.to(to).emit('chat:message', { ...message, notificationId: notification.id });
+          } else {
+            this.io.to(to).emit('chat:message', { ...message });
+          }
+        }
+      });
+
+      socket.on('notification:read', data => this.notificationStore.del(data.userId, data.ids));
 
       /**
        * Destroy session on user logout request