+++ start/routes.ts
@@ -1,0 +1,19 @@
+import './routes/index';
+
+/**
+ * API v1
+ */
+import './routes/apis/v1/auth';
+import './routes/apis/v1/categories';
+import './routes/apis/v1/chat';
+import './routes/apis/v1/courses';
+import './routes/apis/v1/health';
+import './routes/apis/v1/lessons';
+import './routes/apis/v1/me';
+import './routes/apis/v1/stat';
+import './routes/apis/v1/users';
+
+/**
+ * Other routes
+ */
+import './routes/videos';