+++ start/routes/apis/v1/me.ts
@@ -2,6 +2,10 @@ import Route from '@ioc:Adonis/Core/Route';
 
 Route.group(() => {
   Route.get('/', 'Api/v1/MeController.show').middleware('role:admin,teacher,student').as('me.show');
+  Route.get('/messages/:chat_id', 'Api/v1/MeController.chatHistory')
+    .middleware('role:admin,teacher,student')
+    .as('me.chat-history');
+  Route.patch('/info', 'Api/v1/MeController.updateInfo').middleware('role:admin,teacher,student').as('me.update-info');
   Route.patch('/email', 'Api/v1/MeController.changeEmail')
     .middleware('role:admin,teacher,student')
     .as('me.change-email');
@@ -14,4 +18,6 @@ Route.group(() => {
   Route.patch('/contacts', 'Api/v1/MeController.updateContacts')
     .middleware('role:admin,teacher,student')
     .as('me.update-contacts');
-}).prefix('me');
+})
+  .prefix('api/v1/me')
+  .middleware('auth');