+++ start/routes/apis/v1/stat.ts
@@ -2,4 +2,6 @@ import Route from '@ioc:Adonis/Core/Route';
 
 Route.group(() => {
   Route.get('/', 'Api/v1/StatController.index').middleware('role:admin,teacher,student').as('stat.index');
-}).prefix('stat');
+})
+  .prefix('api/v1/stat')
+  .middleware('auth');