+++ CHANGELOG.md
@@ -4,6 +4,22 @@
 
 - API deployment with Docker and Nginx web server.
 
+## [1.2.1] - 2022-03-06
+
+## Added
+
+- Saving latest notifications about new messages
+- Saving recent conversation messages based on constant (now 100)
+- Simple chat implementation between users
+- Added `about` and `last_login` columns for user
+- Updating info api route of authorized user
+
+### Fix
+
+- Refactored routes import
+- Fixed updating user info
+- Cleaning up session store before starting websocket server
+
 ## [1.1.0] - 2022-01-31
 
 ### Added